// React/app/GerAdv_TS/ViaRecebimento/Interfaces/interface.ViaRecebimento.ts
// 100% auto-generated by Source Genesys WebApi Front & Back Creator
                     // copyright © 2000-2025 Menphis - Sistemas Inteligentes
                     // This file is part of the Source Genesys project                     
                    'use client';

export interface IViaRecebimento {
// 202501251
    id: number;
 
	nome: string,
}

export interface IViaRecebimentoFormProps {
  id: number;
  onClose: () => void;
  onSuccess: (registro?: any) => void;
  onError: () => void;
}

export interface IViaRecebimentoIncProps {
    id: number;
    onClose: () => void;
    onSuccess: (registro?: any) => void;
    onError: () => void;
}